Change should to is
Advancement Plaques was causing log spam when it was displaying an advancement. I looked at your code guessed it was might be a conflict with WTHIT and so talked linked this code to Deirn and asked if he saw an issue. This was the change he said was needed. Was originally just going to create an issue, but those seem to be disabled on your repo, so figured I might as well PR the change. We are added the mod to All of Fabric 4 and so figured I could find the log spam issue and get it resolved. Love the mod!